weather.feature scenario ,WeatherResponse, WeatherRequester and WeatherStepDefs completed

+++ src/test/java/model/Sys.java
@@ -2,11 +2,11 @@ package model;
 
 public class Sys {
     private int type;
-    private int id;
+    private long id;
     private double message;
     private String country;
-    private int sunrise;
-    private int sunset;
+    private long sunrise;
+    private long sunset;
 
     public int getType() {
         return type;
@@ -16,11 +16,11 @@ public class Sys {
         this.type = type;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -40,19 +40,19 @@ public class Sys {
         this.country = country;
     }
 
-    public int getSunrise() {
+    public long getSunrise() {
         return sunrise;
     }
 
-    public void setSunrise(int sunrise) {
+    public void setSunrise(long sunrise) {
         this.sunrise = sunrise;
     }
 
-    public int getSunset() {
+    public long getSunset() {
         return sunset;
     }
 
-    public void setSunset(int sunset) {
+    public void setSunset(long sunset) {
         this.sunset = sunset;
     }
 }
